# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BlobLink": {
        "description": "Cross-resource reference to a blob (opaque `id` as returned by the API).",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
...
      "MoveOwner": {
        "additionalProperties": false,
        "description": "Target owning team for a library resource move (`owner` string matches GET responses).",
        "example": {
          "owner": "team_target_id"
        },
        "properties": {
          "owner": {
            "type": "string"
          }
        },
        "required": [
          "owner"
        ],
        "type": "object"
      },
...
      "Problem": {
        "description": "[RFC 7807](https://www.rfc-editor.org/rfc/rfc7807) problem document (`application/problem+json`).\n\nCanonical error body for HTTP 4xx/5xx responses. Extension members include `code`.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code (extension member).",
            "type": "string"
          },
          "detail": {
            "description": "Human-readable explanation.",
            "nullable": true,
            "type": "string"
          },
          "instance": {
            "description": "Optional URI reference that identifies the specific occurrence.",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "description": "HTTP status code.",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "description": "Short, stable summary of the problem class.",
            "type": "string"
          },
          "type": {
            "description": "URI reference identifying the problem type.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "Song": {
        "properties": {
          "blobs": {
            "description": "Linked blob assets (`id` is the blob resource identifier).",
            "items": {
              "$ref": "#/components/schemas/BlobLink"
            },
            "type": "array"
          },
          "data": {
            "$ref": "#/components/schemas/SongDataSchema"
          },
          "id": {
            "type": "string"
          },
          "not_a_song": {
            "description": "When true, this record is not treated as a musical song (e.g. scripture or spoken content).",
            "type": "boolean"
          },
          "owner": {
            "type": "string"
          },
          "user_specific_addons": {
            "$ref": "#/components/schemas/SongUserSpecificAddons"
          }
        },
        "required": [
          "id",
          "owner",
          "not_a_song",
          "blobs",
          "data",
          "user_specific_addons"
        ],
        "type": "object"
      },
...
      "SongUserSpecificAddons": {
        "properties": {
          "liked": {
            "type": "boolean"
          }
        },
        "required": [
          "liked"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "description": "Session cookie returned after a successful authentication flow",
        "in": "cookie",
        "name": "sso_session",
        "type": "apiKey"
      },
      "SessionToken": {
        "description": "Session override using `Authorization: Bearer <session>` header",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Worship Viewer API",
    "version": "2.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/collections/{id}/songs": {
      "get": {
        "operationId": "get_collection_songs",
        "parameters": [
          {
            "description": "Collection identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Page index, zero-based. Omit with `page_size` for full list.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "nullable": true,
              "type": "integer"
            }
          },
          {
            "description": "Items per page. Must be 1–500. Defaults to 50. Omit with `page` for full list.",
            "example": 50,
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 500,
              "minimum": 1,
              "nullable": true,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Song"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Return the songs for a collection. `X-Total-Count` is the total before paging."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid collection identifier or pagination"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Collection not found"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to fetch collection songs"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Collections"
        ]
      }
    },
...
    "/api/v1/setlists/{id}/songs": {
      "get": {
        "operationId": "get_setlist_songs",
        "parameters": [
          {
            "description": "Setlist identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Page index, zero-based. Omit with `page_size` for full list.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "nullable": true,
              "type": "integer"
            }
          },
          {
            "description": "Items per page. Must be 1–500. Defaults to 50. Omit with `page` for full list.",
            "example": 50,
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 500,
              "minimum": 1,
              "nullable": true,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Song"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Return the songs for a setlist. `X-Total-Count` is the total before paging."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid setlist identifier or pagination"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Setlist not found"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to fetch setlist songs"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Setlists"
        ]
      }
    },
...
    "/api/v1/songs/{id}/move": {
      "post": {
        "operationId": "move_song",
        "parameters": [
          {
            "description": "Song identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveOwner"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            },
            "description": "Song moved to the target team, or unchanged when already owned by that team (idempotent). Moving does not add or remove the song from collections."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid `owner` team id"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Song not found, target team not found, or caller lacks library write access on the current or destination team"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to move song"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Songs"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Same origin as the web app (override per deployment).",
      "url": "/"
    },
    {
      "description": "Production deployment (public web app).",
      "url": "https://app.worshipviewer.com"
    }
  ],
...
}